+++ lib/norm.dart
@@ -1,4 +1,4 @@
 /// Vector norm types
 ///
-/// Actually, vector norm defines a metric for considered space
+/// Vector norm defines a metric for considered space
 enum Norm { euclidean, manhattan }