+++ test/float32x4_vector_test.dart
@@ -1,5 +1,6 @@
 import 'dart:typed_data';
 
+import 'package:ml_linalg/distance.dart';
 import 'package:ml_linalg/linalg.dart';
 import 'package:ml_linalg/src/matrix/float32x4_matrix.dart';
 import 'package:ml_linalg/src/vector/float32x4/float32x4_vector.dart';
@@ -373,22 +374,53 @@ void main() {
           reason: 'Wrong vector distance calculation');
     });
 
-    test('should find vectors distance', () {
+    test('should find euclidean distance', () {
       final vector1 = Float32x4Vector.from([10.0, 3.0, 4.0, 7.0, 9.0, 12.0]);
       final vector2 = Float32x4Vector.from([1.0, 3.0, 2.0, 11.5, 10.0, 15.5]);
-      expect(vector1.distanceTo(vector2, Norm.euclidean),
-          equals(10.88577052853862),
-          reason: 'Wrong vector distance calculation');
-      expect(vector1.distanceTo(vector2, Norm.manhattan), equals(20.0),
-          reason: 'Wrong vector distance calculation');
+      expect(vector1.distanceTo(vector2, distance: Distance.euclidean),
+          equals(10.88577052853862));
+    });
+
+    test('should find manhattan distance', () {
+      final vector1 = Float32x4Vector.from([10.0, 3.0, 4.0, 7.0, 9.0, 12.0]);
+      final vector2 = Float32x4Vector.from([1.0, 3.0, 2.0, 11.5, 10.0, 15.5]);
+      expect(vector1.distanceTo(vector2, distance: Distance.manhattan),
+          equals(20.0));
+    });
+
+    test('should find cosine distance (the same vectors)', () {
+      final vector1 = Float32x4Vector.from([1.0, 0.0]);
+      final vector2 = Float32x4Vector.from([1.0, 0.0]);
+      expect(vector1.distanceTo(vector2, distance: Distance.cosine),
+          equals(0.0));
+    });
+
+    test('should find cosine distance (different vectors)', () {
+      final vector1 = Float32x4Vector.from([4.0, 3.0]);
+      final vector2 = Float32x4Vector.from([2.0, 4.0]);
+      expect(vector1.distanceTo(vector2, distance: Distance.cosine),
+          closeTo(0.1055, 1e-4));
+    });
+
+    test('should find cosine distance (different vectors with negative '
+        'elements)', () {
+      final vector1 = Float32x4Vector.from([4.0, -3.0]);
+      final vector2 = Float32x4Vector.from([-2.0, 4.0]);
+      expect(vector1.distanceTo(vector2, distance: Distance.cosine),
+          closeTo(1.8944, 1e-4));
+    });
+
+    test('should find cosine distance (one of two vectors is zero-vector)', () {
+      final vector1 = Float32x4Vector.from([0.0, 0.0]);
+      final vector2 = Float32x4Vector.from([-2.0, 4.0]);
+      expect(() => vector1.distanceTo(vector2, distance: Distance.cosine),
+          throwsException);
     });
 
     test('should find vector norm', () {
       final vector = Float32x4Vector.from([1.0, 2.0, 3.0, 4.0, 5.0]);
-      expect(vector.norm(Norm.euclidean), equals(closeTo(7.41, 1e-2)),
-          reason: 'Wrong norm calculation');
-      expect(vector.norm(Norm.manhattan), equals(15.0),
-          reason: 'Wrong norm calculation');
+      expect(vector.norm(Norm.euclidean), equals(closeTo(7.41, 1e-2)));
+      expect(vector.norm(Norm.manhattan), equals(15.0));
     });
 
     test('should normalize itself (eucleadean norm)', () {